refactor: Like 공통 클래스 패키지 리팩토링
