+++ src/main/java/com/mobble/mobbleserver/domain/like/articleLike/service/ArticleLikeService.java
@@ -6,8 +6,8 @@ import com.mobble.mobbleserver.domain.clubMember.validator.ClubMemberValidator;
 import com.mobble.mobbleserver.domain.like.articleLike.dto.response.ArticleLikeMemberListResponseDto;
 import com.mobble.mobbleserver.domain.like.articleLike.entity.ArticleLike;
 import com.mobble.mobbleserver.domain.like.articleLike.repository.ArticleLikeRepository;
-import com.mobble.mobbleserver.domain.like.entity.LikeType;
-import com.mobble.mobbleserver.domain.like.service.AbstractLikeService;
+import com.mobble.mobbleserver.domain.like.baseLike.entity.LikeType;
+import com.mobble.mobbleserver.domain.like.baseLike.service.AbstractLikeService;
 import com.mobble.mobbleserver.domain.member.entity.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
