+++ src/main/java/com/mobble/mobbleserver/domain/like/baseLike/controller/LikeController.java
@@ -1,9 +1,9 @@
-package com.mobble.mobbleserver.domain.like.controller;
+package com.mobble.mobbleserver.domain.like.baseLike.controller;
 
 import com.mobble.mobbleserver.domain.like.articleLike.service.ArticleLikeService;
-import com.mobble.mobbleserver.domain.like.dto.response.LikeToggleResponseDto;
-import com.mobble.mobbleserver.domain.like.entity.LikeType;
-import com.mobble.mobbleserver.domain.like.service.LikeDispatcherService;
+import com.mobble.mobbleserver.domain.like.baseLike.dto.response.LikeToggleResponseDto;
+import com.mobble.mobbleserver.domain.like.baseLike.entity.LikeType;
+import com.mobble.mobbleserver.domain.like.baseLike.service.LikeDispatcherService;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.like.LikeErrorCode;
 import lombok.RequiredArgsConstructor;
