+++ src/main/java/com/mobble/mobbleserver/domain/like/baseLike/entity/BaseLike.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.domain.like.entity;
+package com.mobble.mobbleserver.domain.like.baseLike.entity;
 
 import com.mobble.mobbleserver.domain.member.entity.Member;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
