+++ src/main/java/com/mobble/mobbleserver/domain/like/baseLike/entity/LikeType.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.domain.like.entity;
+package com.mobble.mobbleserver.domain.like.baseLike.entity;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
