+++ src/main/java/com/mobble/mobbleserver/domain/like/clubLike/entity/ClubLike.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.domain.like.clubLike.entity;
 
 import com.mobble.mobbleserver.domain.club.club.entity.Club;
-import com.mobble.mobbleserver.domain.like.entity.BaseLike;
+import com.mobble.mobbleserver.domain.like.baseLike.entity.BaseLike;
 import com.mobble.mobbleserver.domain.member.entity.Member;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.like.LikeErrorCode;
