+++ src/main/java/com/mobble/mobbleserver/domain/like/commentLike/service/CommentLikeService.java
@@ -5,8 +5,8 @@ import com.mobble.mobbleserver.domain.comment.entity.Comment;
 import com.mobble.mobbleserver.domain.comment.validator.CommentValidator;
 import com.mobble.mobbleserver.domain.like.commentLike.entity.CommentLike;
 import com.mobble.mobbleserver.domain.like.commentLike.repository.CommentLikeRepository;
-import com.mobble.mobbleserver.domain.like.entity.LikeType;
-import com.mobble.mobbleserver.domain.like.service.AbstractLikeService;
+import com.mobble.mobbleserver.domain.like.baseLike.entity.LikeType;
+import com.mobble.mobbleserver.domain.like.baseLike.service.AbstractLikeService;
 import com.mobble.mobbleserver.domain.member.entity.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
